update gitlab scripts

+++ scripts/local-gitlab/run-ee.sh
@@ -12,20 +12,32 @@
 GITLAB_IMAGE_TAG=$1
 GITLAB_HOME=$2
 GITLAB_HOST=$3
+PERSIST_GITLAB_LOCALLY=$4
 
 echo ""
 echo "Starting fresh install of GitLab Enterprise Edition, using tag: ${GITLAB_IMAGE_TAG}"
-echo "Removing any previous install at path: ${GITLAB_HOME}"
+
+
+if [ "${PERSIST_GITLAB_LOCALLY}" == "true" ];then
+  echo "Removing any previous install at path: ${GITLAB_HOME}"
+  echo ""
+
+  rm -rf "${GITLAB_HOME}"
+
+  docker run \
+    -d=true \
+    --hostname "${GITLAB_HOST}" \
+    --publish 443:443 --publish 80:80 --publish 22:22 \
+    --name gitlab \
+    gitlab/gitlab-ee:"${GITLAB_IMAGE_TAG}"
+else
+
+  docker run \
+    -d=true \
+    --hostname "${GITLAB_HOST}" \
+    --publish 443:443 --publish 80:80 --publish 22:22 \
+    --name gitlab \
+    gitlab/gitlab-ee:"${GITLAB_IMAGE_TAG}"
+fi
+
 echo ""
-
-rm -rf "${GITLAB_HOME}"
-
-docker run \
-  -d=true \
-  --hostname "${GITLAB_HOST}" \
-  --publish 443:443 --publish 80:80 --publish 22:22 \
-  --name gitlab \
-  --volume "${GITLAB_HOME}"/config:/etc/gitlab \
-  --volume "${GITLAB_HOME}"/logs:/var/log/gitlab \
-  --volume "${GITLAB_HOME}"/data:/var/opt/gitlab \
-  gitlab/gitlab-ee:"${GITLAB_IMAGE_TAG}"
